+++ src/Components/Project/ProjectItem.tsx
@@ -1,25 +1,6 @@
-import { FaReact, FaJava, FaHtml5, FaGithub } from "react-icons/fa";
-import { IoLogoCss3, IoLogoJavascript } from "react-icons/io";
-import { SiTailwindcss, SiMysql, SiSpring } from "react-icons/si";
-import { DiPostgresql } from "react-icons/di";
+import Icon from "../Icons/Icons";
 
 import { projectItemProps } from "./interface/projectItemProps";
-
-const iconStyling = "fill-white mr-2 h-[32px] w-[32px]";
-
-const techIcons = {
-  react: <FaReact className={iconStyling} />,
-  css: <IoLogoCss3 className={iconStyling} />,
-  javascript: <IoLogoJavascript className={iconStyling} />,
-  tailwind: <SiTailwindcss className={iconStyling} />,
-  java: <FaJava className={iconStyling} />,
-  mysql: <SiMysql className={iconStyling} />,
-  postgres: <DiPostgresql className={iconStyling} />,
-  html: <FaHtml5 className={iconStyling} />,
-  git: <FaGithub className={iconStyling} />,
-  github: <FaGithub className={iconStyling} />,
-  spring: <SiSpring className={iconStyling} />,
-};
 
 const ProjectItem = (props: projectItemProps) => {
   return (
@@ -29,7 +10,7 @@ const ProjectItem = (props: projectItemProps) => {
       <div className="hover:opacity-100 bg-dark absolute top-0 left-0 w-full h-full flex flex-col items-center justify-center opacity-0 transition-all ease-in-out duration-300">
         <div className="flex mb-2">
 
-          {props.technologies.map((tech) => {
+        {props.technologies.map((tech) => {
             // Extract the technology name from the image path
             const techName = tech.split("/").pop().split(".")[0];
             return (
@@ -38,7 +19,7 @@ const ProjectItem = (props: projectItemProps) => {
                 className="saturate-0 mr-2"
                 style={{ color: "white", width: "32px", height: "32px" }}
               >
-                {techIcons[techName]}
+                <Icon name={techName} />
               </span>
             );
           })}
